trade-from-back
added in a trade from back function that looks for cards to trade from
back to front. It just takes the last value from the trade from front
function + 2 (or 1 if we want to keep the first card in our hand) and
keeps looking back from there. I gave them separate lists so we can
value them differently